% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{add_agg_flags}
\alias{add_agg_flags}
\title{Add aggregation level flags}
\usage{
add_agg_flags(df)
}
\arguments{
\item{df}{Processed data frame}
}
\value{
Data frame with aggregation flags
}
\description{
Determines whether each row is state, district, or school level data
and adds NYC and charter school flags.
}
\keyword{internal}
