% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_everything}
\alias{aggregate_everything}
\title{Makes all-grade and sub-grade aggregates from a clean data file}
\usage{
aggregate_everything(clean_df, verbose = TRUE)
}
\arguments{
\item{clean_df}{output of clean_assess_db or fetch_assess_db}

\item{verbose}{logical, print status updates to the console?}
}
\value{
data.frame, with all-grade and sub-grade aggregations
}
\description{
Makes all-grade and sub-grade aggregates from a clean data file
}

