% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_grades}
\alias{aggregate_grades}
\title{Aggregate results by subject across grade levels}
\usage{
aggregate_grades(df)
}
\arguments{
\item{df}{data.frame, output of clean_assess_db or fetch_assess_db}
}
\value{
data.frame with *only* the sch/dist aggregate scores.
}
\description{
Aggregate results by subject across grade levels
}

