% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_db.R
\name{assess_db}
\alias{assess_db}
\title{Create a assess_db object}
\usage{
assess_db(test_years, verbose = TRUE)
}
\arguments{
\item{test_years}{numeric vector of test years}

\item{verbose}{should assess_db print status updates as
it generates the object?  default is TRUE.}
}
\examples{
\dontrun{
ex_assess_db <- assess_db(
  test_years = c(2014:2016)
 )

is.assess_db(ex_assess_db)
print(ex_assess_db)
}
}

