% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_growth.R
\name{assess_growth}
\alias{assess_growth}
\title{Assessment Growth (Same Grade)}
\usage{
assess_growth(clean_assess_start, clean_assess_end)
}
\arguments{
\item{clean_assess_start}{data.frame, start of window.
output of clean_assess_db}

\item{clean_assess_end}{data.frame, end of window.
output of clean_assess_db}
}
\value{
data.frame
}
\description{
change in assessment performance,
same grade, across two test years.
}

