% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{build_src_url}
\alias{build_src_url}
\title{Build NYSED Report Card Database URL}
\usage{
build_src_url(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Full download URL
}
\description{
Constructs the download URL based on year. URL patterns changed over time:
\itemize{
\item 2018+: /files/essa/YY-YY/SRCYYYY.zip
\item 2014-2017: /files/reportcards/YY-YY/SRCYYYY.zip
}
}
\keyword{internal}
