% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{check_mdbtools}
\alias{check_mdbtools}
\title{Check if mdbtools is available}
\usage{
check_mdbtools()
}
\value{
Logical indicating if mdb-export command is available
}
\description{
Check if mdbtools is available
}
\keyword{internal}
