% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_db.R
\name{clean_assess_db}
\alias{clean_assess_db}
\title{Clean an assessment database file}
\usage{
clean_assess_db(df, test_year, suppressed_as_NA = TRUE,
  cohort_kind = "college_entry")
}
\arguments{
\item{df}{data.frame (output of `get_raw_assess_db`)}

\item{test_year}{testing year}

\item{suppressed_as_NA}{should records supressed for small n
(recorded as '-' in the raw file) be converted to NA?  default is TRUE}

\item{cohort_kind}{should we name cohorts by their college entry
or college graduation year?  default is college_entry.}
}
\value{
tbl_df
}
\description{
Clean an assessment database file
}

