% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_growth.R
\name{cohort_growth}
\alias{cohort_growth}
\title{Cohort Assessment Growth}
\usage{
cohort_growth(clean_assess_start, clean_assess_end)
}
\arguments{
\item{clean_assess_start}{data.frame, start of window.
output of clean_assess_db}

\item{clean_assess_end}{data.frame, end of window.
output of clean_assess_db}
}
\value{
data.frame
}
\description{
change in assessment performance following an
implicit cohort over multiple years
}

