% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{custom_aggregate}
\alias{custom_aggregate}
\title{make a custom aggregation}
\usage{
custom_aggregate(assess_df, bedscode, grades, cust_suffix = "_custom")
}
\arguments{
\item{assess_df}{data.frame, output of fetch_aggregate_percentile_assess_db}

\item{bedscode}{bedscode for the school you want to custom aggregate}

\item{grades}{grades that are in this custom aggregation}

\item{cust_suffix}{suffix to throw onto the bedscode}
}
\value{
data.frame
}
\description{
some charters or schools have multiple 'campuses' inside of the same
bedscode. make a custom aggregation by grade level.
}

