% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\docType{data}
\name{DATA_CACHE_DIR}
\alias{DATA_CACHE_DIR}
\title{Data cache directory for converted CSV files}
\format{
An object of class \code{character} of length 1.
}
\usage{
DATA_CACHE_DIR
}
\description{
Data cache directory for converted CSV files
}
\keyword{internal}
