% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graduation.R
\docType{data}
\name{DEFAULT_COHORT_TYPE}
\alias{DEFAULT_COHORT_TYPE}
\title{Default cohort type for graduation rates}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_COHORT_TYPE
}
\description{
Default cohort type for graduation rates
}
\keyword{internal}
