% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_src_database}
\alias{download_src_database}
\title{Download NYSED Report Card Database}
\usage{
download_src_database(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Path to the extracted .mdb file
}
\description{
Downloads and extracts the Access database file from NYSED.
}
\keyword{internal}
