% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_enrollment.R
\name{enr_grade_aggs}
\alias{enr_grade_aggs}
\title{Custom Enrollment Grade Level Aggregates}
\usage{
enr_grade_aggs(df)
}
\arguments{
\item{df}{A tidy enrollment df (output of tidy_enr)}
}
\value{
df of aggregated enrollment data
}
\description{
Creates aggregations for common grade groupings: K-8, 9-12 (HS), K-12.
}
