% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{extract_district_name}
\alias{extract_district_name}
\title{Extract district name from data}
\usage{
extract_district_name(raw, district_ids)
}
\arguments{
\item{raw}{Raw data frame}

\item{district_ids}{District BEDS codes}
}
\value{
District names
}
\description{
Extract district name from data
}
\keyword{internal}
