% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract_mdb}
\alias{extract_mdb}
\title{Finds .mdb files in a temp directory and processes as data frames}
\usage{
extract_mdb(file_list)
}
\arguments{
\item{file_list}{list, output of zip_to_temp}
}
\value{
a list of data frames
}
\description{
Finds .mdb files in a temp directory and processes as data frames
}

