% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_db.R
\name{fetch_and_aggregate_assess_db}
\alias{fetch_and_aggregate_assess_db}
\title{Fetch NY State assessment db, and calculate multi-grade aggregates}
\usage{
fetch_and_aggregate_assess_db(test_year, verbose = TRUE)
}
\arguments{
\item{test_year}{4 digit integer representing the end of the desired school
year.  eg, 2014-15 is 2015.}

\item{verbose}{logical, print status updates to the console?}
}
\value{
data.frame
}
\description{
Fetch NY State assessment db, and calculate multi-grade aggregates
}

