% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_enrollment.R
\name{fetch_enr_nyc}
\alias{fetch_enr_nyc}
\title{Get NYC enrollment data}
\usage{
fetch_enr_nyc(end_year, level = "school", tidy = TRUE, use_cache = TRUE)
}
\arguments{
\item{end_year}{School year end (or vector of years)}

\item{level}{Data level: "school" or "district"}

\item{tidy}{If TRUE, returns tidy format}

\item{use_cache}{If TRUE, uses cache}
}
\value{
Data frame with NYC enrollment data
}
\description{
Convenience function to get enrollment data for NYC DOE schools only.
}
\examples{
\dontrun{
# Get NYC school enrollment
nyc_enr <- fetch_enr_nyc(2024)
}
}
