% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_enrollment.R
\name{filter_grade_span}
\alias{filter_grade_span}
\title{Filter enrollment data by grade span}
\usage{
filter_grade_span(df, span = "k12")
}
\arguments{
\item{df}{Tidy enrollment data frame}

\item{span}{Grade span: "pk12", "k12", "k8", "hs", "elem" (K-5), "middle" (6-8)}
}
\value{
Filtered data frame
}
\description{
Convenience function to filter to common grade spans.
}
