% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{full_aggregation_scaffold}
\alias{full_aggregation_scaffold}
\title{Create the full aggregation scaffold for all records in a clean assess df}
\usage{
full_aggregation_scaffold(clean_df)
}
\arguments{
\item{clean_df}{output of clean_assess_db or fetch_assess_db}
}
\value{
data.frame ready to join back to the clean_df
}
\description{
Create the full aggregation scaffold for all records in a clean assess df
}

