% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_db.R
\name{get_all_raw_rc}
\alias{get_all_raw_rc}
\title{Reads in all the raw report card databases}
\usage{
get_all_raw_rc()
}
\value{
list of report card db lists
}
\description{
Reads in all the raw report card databases
}

