% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get Available Years}
\usage{
get_available_years()
}
\value{
Named list with min_year and max_year
}
\description{
Returns the range of years for which enrollment data is available.
}
\examples{
get_available_years()
}
