% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the cache directory for nyschooldata}
\usage{
get_cache_dir()
}
\value{
Path to cache directory
}
\description{
Returns the path to the cache directory, creating it if necessary.
Uses rappdirs for cross-platform cache location.
}
\keyword{internal}
