% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_path}
\alias{get_cache_path}
\title{Get cached file path for a given year}
\usage{
get_cache_path(end_year, type = "raw")
}
\arguments{
\item{end_year}{School year end}

\item{type}{File type ("raw" or "processed" or "tidy")}
}
\value{
Path to cached file
}
\description{
Get cached file path for a given year
}
\keyword{internal}
