% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{get_grad_cache_path}
\alias{get_grad_cache_path}
\title{Get cache path for converted CSV file}
\usage{
get_grad_cache_path(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Path to cached CSV file
}
\description{
Get cache path for converted CSV file
}
\keyword{internal}
