% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_db.R
\name{get_raw_assess_db}
\alias{get_raw_assess_db}
\title{Read in the raw assessment database from the NYSED website}
\usage{
get_raw_assess_db(test_year)
}
\arguments{
\item{test_year}{4 digit integer representing the end of the desired school
year.  eg, 2014-15 is 2015.}
}
\value{
data frame with assessment results, by school
}
\description{
Read in the raw assessment database from the NYSED website
}

