% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_archive}
\alias{get_raw_enr_archive}
\title{Download archive format (1977-2011) NYSED data}
\usage{
get_raw_enr_archive(end_year, level = "district")
}
\arguments{
\item{end_year}{School year end}

\item{level}{Data level}
}
\value{
Raw data frame
}
\description{
Download archive format (1977-2011) NYSED data
}
\keyword{internal}
