% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_demographics}
\alias{get_raw_enr_demographics}
\title{Download demographics enrollment data from NYSED}
\usage{
get_raw_enr_demographics(end_year, level = "school", category = "race")
}
\arguments{
\item{end_year}{School year end}

\item{level}{Data level: "school" or "district"}

\item{category}{Demographic category: "race", "economic", "ell", "swd"}
}
\value{
Raw data frame with demographic breakdowns
}
\description{
Downloads enrollment broken down by demographic subgroups (race/ethnicity,
economically disadvantaged, ELL, SWD).
}
\keyword{internal}
