% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_legacy}
\alias{get_raw_enr_legacy}
\title{Download legacy format (2012-2021) NYSED data}
\usage{
get_raw_enr_legacy(end_year, level = "school")
}
\arguments{
\item{end_year}{School year end}

\item{level}{Data level}
}
\value{
Raw data frame
}
\description{
Download legacy format (2012-2021) NYSED data
}
\keyword{internal}
