% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_modern}
\alias{get_raw_enr_modern}
\title{Download modern format (2022+) NYSED data}
\usage{
get_raw_enr_modern(end_year, level = "school")
}
\arguments{
\item{end_year}{School year end}

\item{level}{Data level}
}
\value{
Raw data frame
}
\description{
Download modern format (2022+) NYSED data
}
\keyword{internal}
