% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_db.R
\name{get_raw_rc_db}
\alias{get_raw_rc_db}
\title{Read in the raw report card database from the NYSED website}
\usage{
get_raw_rc_db(end_year)
}
\arguments{
\item{end_year}{4 digit integer representing the end of the desired school
year.  eg, 2014-15 is 2015.}
}
\value{
list of data frames for all the tables in the report card db.
}
\description{
Read in the raw report card database from the NYSED website
}

