% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{grade_runs}
\alias{grade_runs}
\title{Finds grade runs}
\usage{
grade_runs(min_gr, max_gr)
}
\arguments{
\item{min_gr}{smallest grade}

\item{max_gr}{largest grade}
}
\value{
list with grade runs
}
\description{
Finds grade runs
}

