% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
left \%||\% right
}
\arguments{
\item{left}{Left value}

\item{right}{Right value (default)}
}
\value{
left if not null/NA, else right
}
\description{
Returns left side if not null, otherwise right side
}
\keyword{internal}
