% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedscode.R
\name{interpret_bedscode}
\alias{interpret_bedscode}
\title{Interpret bedscode}
\usage{
interpret_bedscode(df)
}
\arguments{
\item{df}{data frame containing the column 'bedscode'
ie, output of clean_assess_db}
}
\value{
data.frame
}
\description{
breaks out bedscode into element
}

