% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedscode.R
\name{interpret_bedscode_complete}
\alias{interpret_bedscode_complete}
\title{Interpret bedscode (complete)}
\usage{
interpret_bedscode_complete(df)
}
\arguments{
\item{df}{data frame containing the column 'bedscode'
ie, output of clean_assess_db}
}
\value{
data.frame
}
\description{
Interpret the bedscode, and return everything
}

