% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyschooldata-package.R
\docType{package}
\name{nyschooldata-package}
\alias{nyschooldata}
\alias{nyschooldata-package}
\title{nyschooldata: Fetch and Process New York School Data}
\description{
Downloads and processes school data from the New York State Education
Department (NYSED). Provides functions for fetching enrollment data and
transforming it into tidy format for analysis.
}
\section{Main Functions}{

\describe{
\item{\code{\link{fetch_enr}}}{Fetch enrollment data for a school year}
\item{\code{\link{fetch_enr_years}}}{Fetch enrollment data for multiple years}
\item{\code{\link{fetch_enr_school}}}{Fetch enrollment for a specific school by BEDS code}
\item{\code{\link{fetch_enr_district}}}{Fetch enrollment for a specific district}
\item{\code{\link{fetch_enr_nyc}}}{Fetch NYC DOE enrollment data}
\item{\code{\link{tidy_enr}}}{Transform wide enrollment data to long format}
\item{\code{\link{id_enr_aggs}}}{Identify aggregation levels (state/district/school)}
\item{\code{\link{enr_grade_aggs}}}{Create grade-level aggregates (K-8, HS, K-12)}
}
}

\section{Data Sources}{

Data is downloaded from the NYSED data portal at
\url{https://data.nysed.gov/}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://almartin82.github.io/nyschooldata/}
  \item \url{https://github.com/almartin82/nyschooldata}
  \item Report bugs at \url{https://github.com/almartin82/nyschooldata/issues}
}

}
\author{
\strong{Maintainer}: Andy Martin \email{almartin@gmail.com}

}
\keyword{internal}
