% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{nysd_clean_names}
\alias{nysd_clean_names}
\title{NYSD clean names}
\usage{
nysd_clean_names(name_vector)
}
\arguments{
\item{name_vector}{character vector of school names}
}
\value{
character vector of codes
}
\description{
NYSD clean names
}

