% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\docType{data}
\name{NYSED_GRAD_BASE_URL}
\alias{NYSED_GRAD_BASE_URL}
\title{Base URL for NYSED graduation rate downloads}
\format{
An object of class \code{character} of length 1.
}
\usage{
NYSED_GRAD_BASE_URL
}
\description{
Base URL for NYSED graduation rate downloads
}
\keyword{internal}
