% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proficiency_hist.R
\name{p_proficiency_hist_single_subj_sch_comparison}
\alias{p_proficiency_hist_single_subj_sch_comparison}
\title{Proficiency History, One Subject, Comparing Two Entities}
\usage{
p_proficiency_hist_single_subj_sch_comparison(assess_db_all, bedscodes,
  subjects, grades, subgroup_codes = "01")
}
\arguments{
\item{assess_db_all}{data.frame, multiple years of assessment history}

\item{bedscodes}{character vector, one or more bedscode(s)}

\item{subjects}{character vector, one or more of c('ELA', 'Math')}

\item{grades}{vector, one or more grades, inclusive of 'All'.  c('All', 3, 4)}

\item{subgroup_codes}{vector, one or more subgroup codes}
}
\value{
ggplot
}
\description{
exported visualization that compares the proficiency
history of two NY state entities - ie a school and the state.
}

