% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{parse_beds_code}
\alias{parse_beds_code}
\title{Parse BEDS code into components}
\usage{
parse_beds_code(beds_code)
}
\arguments{
\item{beds_code}{Character vector of BEDS codes}
}
\value{
Data frame with parsed components
}
\description{
BEDS codes are 12 digits: DDDDDDSSSSCC
\itemize{
\item DDDDDD: District code (6 digits)
\item SSSS: School code (4 digits)
\item CC: Check digits (2 digits)
}
}
\examples{
parse_beds_code("010100010018")
# Returns: district_code = "010100", school_code = "0001", check_digits = "18"
}
