% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{parse_entity_codes}
\alias{parse_entity_codes}
\title{Parse entity codes to identify entity type}
\usage{
parse_entity_codes(df)
}
\arguments{
\item{df}{Data frame with entity_cd column}
}
\value{
Data frame with parsed entity information
}
\description{
NYSED entity codes follow patterns that indicate the type of entity:
\itemize{
\item State: "000000000000" or starts with "0000000000"
\item County: "000XXX000000" (XXX = county code)
\item Need/Resource Category: "0000000000XX"
\item District: "XXXXXXXXXXXX" (12 digits, ends in 0000)
\item School: "XXXXXXXXXXXX" (12 digits, doesn't end in 0000)
}
}
\keyword{internal}
