% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attainment.R
\name{prof_attainment_percentile}
\alias{prof_attainment_percentile}
\title{Proficinency Attainment Percentile}
\usage{
prof_attainment_percentile(percent_prof, comparison_df, school_logical,
  district_logical, multigrade_logical, subject, grade = NA, subgroup = 1)
}
\arguments{
\item{percent_prof}{numeric, percent proficient, 0-100 scale}

\item{comparison_df}{data.frame (usually same year) to compare to}

\item{school_logical}{logical, is this a school?  schools get}

\item{district_logical}{logical, is this a district?}

\item{multigrade_logical}{logical, is this multigrade?}

\item{subject}{character, ELA or Math?}

\item{grade}{grade level?}

\item{subgroup}{character, subgroup number}
}
\value{
numeric length 1 vector
}
\description{
compares schools to schools, districts to districts, to
get a statewide peer attainment percentile.
}

