% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proficiency_hist.R
\name{prof_history_data_prep}
\alias{prof_history_data_prep}
\title{internal data helper function for comparison plots.  D.R.Y!}
\usage{
prof_history_data_prep(assess_db_all, bedscodes, subjects, grades,
  subgroup_codes)
}
\arguments{
\item{assess_db_all}{data.frame, multiple years of assessment history}

\item{bedscodes}{character vector, one or more bedscode(s)}

\item{subjects}{character vector, one or more of c('ELA', 'Math')}

\item{grades}{vector, one or more grades, inclusive of 'All'.  c('All', 3, 4)}

\item{subgroup_codes}{vector, one or more subgroup codes}
}
\value{
data.frame
}
\description{
internal data helper function for comparison plots.  D.R.Y!
}

