% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{read_cache}
\alias{read_cache}
\title{Read data from cache}
\usage{
read_cache(end_year, type = "raw")
}
\arguments{
\item{end_year}{School year end}

\item{type}{File type}
}
\value{
Cached data frame
}
\description{
Read data from cache
}
\keyword{internal}
