% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attainment.R
\name{scale_attainment_percentile}
\alias{scale_attainment_percentile}
\title{Mean Scale Score Attainment Percentile}
\usage{
scale_attainment_percentile(scale_score, comparison_df, school_logical,
  district_logical, multigrade_logical, subject, grade = NA, subgroup = 1)
}
\arguments{
\item{comparison_df}{data.frame (usually same year) to compare to}

\item{school_logical}{logical, is this a school?  schools get}

\item{district_logical}{logical, is this a district?}

\item{multigrade_logical}{logical, is this multigrade?}

\item{subject}{character, ELA or Math?}

\item{grade}{grade level?}

\item{subgroup}{character, subgroup number}
}
\value{
numeric vector length 1
}
\description{
Mean Scale Score Attainment Percentile
}

