% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{sch_aggregation_scaffold}
\alias{sch_aggregation_scaffold}
\title{Creates an aggregation scaffold for one school/district}
\usage{
sch_aggregation_scaffold(bedscode, min_gr, max_gr)
}
\arguments{
\item{bedscode}{character, BEDS identifier}

\item{min_gr}{integer, smallest grade level at this school}

\item{max_gr}{integer, largest grade level at this school}
}
\value{
data.frame, with all of the consecutive sub-grade runs
}
\description{
Creates an aggregation scaffold for one school/district
}

