% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_efficacy.R
\name{subgroup_efficacy}
\alias{subgroup_efficacy}
\title{Subgroup Efficacy}
\usage{
subgroup_efficacy(df, bedscode, year, subject, subgroups = c("01", "02", "03",
  "05", "06", "07", "08", "10", "11", "13", "15", "16"),
  sch_aggregates = TRUE)
}
\arguments{
\item{df}{data.frame, with one grade level or
school level aggregates}

\item{bedscode}{target bedscode}

\item{year}{school year}

\item{subject}{subject}

\item{subgroups}{vector of subgroup codes}

\item{sch_aggregates}{use school level aggregates?}
}
\value{
ggplot
}
\description{
Subgroup Efficacy
}

