% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_enrollment.R
\name{tidy_enr}
\alias{tidy_enr}
\title{Tidy enrollment data}
\usage{
tidy_enr(df)
}
\arguments{
\item{df}{A wide data.frame of processed enrollment data}
}
\value{
A long data.frame of tidied enrollment data
}
\description{
Transforms wide enrollment data to long format with grade_level column.
Each row represents enrollment for one school/district, one grade level.
}
