% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_cache}
\alias{write_cache}
\title{Write data to cache}
\usage{
write_cache(data, end_year, type = "raw")
}
\arguments{
\item{data}{Data frame to cache}

\item{end_year}{School year end}

\item{type}{File type}
}
\description{
Write data to cache
}
\keyword{internal}
