% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{zip_to_temp}
\alias{zip_to_temp}
\title{Download and unzip a file from the web}
\usage{
zip_to_temp(url)
}
\arguments{
\item{url}{url to a zip file}
}
\value{
list, with the temp dir and a data frame of file names from zip file.
}
\description{
Download and unzip a file from the web
}

