% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R, R/get_raw_graduation.R
\name{check_mdbtools}
\alias{check_mdbtools}
\title{Check if mdbtools is available}
\usage{
check_mdbtools()

check_mdbtools()
}
\value{
TRUE if available, stops with error if not

Logical indicating if mdb-export command is available
}
\description{
Verifies that mdb-export is installed and accessible.
}
\keyword{internal}
