% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_db.R
\name{fetch_assess_db}
\alias{fetch_assess_db}
\title{fetch NY State assessment db}
\usage{
fetch_assess_db(end_year)
}
\arguments{
\item{end_year}{}
}
\value{
clean data frame with assessment data
}
\description{
wrapper around get_raw_assess_db and clean_assess_db
}

